[add]修改字段类型为Long

+++ src/main/java/com/wp/community/controller/PublishController.java
@@ -35,7 +35,7 @@ public class PublishController {
 
     //编辑问题
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id, Model model){
+    public String edit(@PathVariable(name = "id") Long id, Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
         model.addAttribute("description",question.getDescription());
@@ -51,7 +51,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request, Model model){
 
         model.addAttribute("title",title);
